requirements.txt
python-dotenv
openai
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pytz
phue
spotipy
discord.py